Make Discourse descriptions more concise
The Discourse node descriptions are quite cluttered, and e.g. we wind up
with silly descriptions like "like on post post #3".

This cleans it up; for post descriptions, we omit the word "post" and
let that be implicit. For like descriptions, we also omit that it was a
post that was liked. I think this will be cleaner.

Test plan: just a change to some strings, snapshots/tests updated.