implement grain transfer interface
this interface allows for grain transfers between user identities. There
are validations in place to ensure an amount exceeding the senders
balance cannot be submitted.

Currently there is no completely lossless library function for
converting a stringified float to grain. This is needed to maintain
fidelity between the frontend and the backend and is currently
WIP (#2034)

test-plan: send grain between accounts. try to send a grain amount
exceeding the sender's balance. Ensure the entered value is stored
correctly in the ledger.json file.