Change initiative directory config to load from plugin context instead of ENV
Now that plugins have their own directories for config, we can use that for the initiatives directory as well instead
of setting it with environment variables. Currently it requires users to put their initiative files in a 'initiatives'
folder beside the config.json.

Test Plan: Place some initiative files in a directory named `initiatives` beside the config.json for the initiatives plugin
and make sure the initiatives are loaded properly into the graph.