Refactor Discourse reference edge creation
In the discourse `createGraph` module, we have a general pattern where
helper functions are responsible for constructing all of the Nodes and
Edges in the graph. This pattern is broken for the reference edges. This
cleanup commit factors the reference edge creation out to follow the
same pattern.

This will enable further cleanups of the discourse createGraph module.

Test plan: Super straightforward refactor, no changes to tests needed.
`yarn test --full` passing implies that real generated Graphs for our
example instance are unchanged.