refactor Discourse graph creation
This commit re-writes Discourse Graph creation so that it proceeds in
two basic steps: first, a data model of the graph is constructed, and
second, that data model is transformed into a WeightedGraph.

This allow us to decouple a lot of complicated logic in the createGraph
method into a data transformer that isn't constructing a mutable object,
which will make it easier to create test cases, etc.

For now, I've just left the existing tests, which have good coverage.
This is mostly to make future changes (e.g. #2029) easier to reason
about.

Test plan: `yarn test`