Implement Initiatives plugin for new CLI (#2027)
* Use initiatives parser for loading initiative directory

Updates the logic in initiativesDirectory.js to use the new initiatives parser when loading initiative files

Test Plan: Ensure initiative files are being parsed and loaded correctly

* Implement Initiatives plugin for new CLI

Sets up a cli2 plugin for the Initiatives plugin and registers it in the bundled plugins

Test Plan: Create an initiatives directory with some example initiatives in the example instance and check if they are successfully loaded into the graph

* Change initiative directory config to load from plugin context instead of ENV

Now that plugins have their own directories for config, we can use that for the initiatives directory as well instead
of setting it with environment variables. Currently it requires users to put their initiative files in a 'initiatives'
folder beside the config.json.

Test Plan: Place some initiative files in a directory named `initiatives` beside the config.json for the initiatives plugin
and make sure the initiatives are loaded properly into the graph.