Temporarily fix #1807 for the initiatives plugin
This commit implements a temporary/hacky fix to close #1807 for the
Initiatives plugin. The basic problem is that the initiatives plugin was
written before the Ledger, and designed to be updated by hand, so we
reference users via URL (e.g. GitHub account url). However, if that user
hasn't interacted with the project's GitHub repos, they won't be in the
Graph and the reference will fail.

This commit adds support for referencing identities in the Ledger by
writing their identity name, e.g. referencing an Identity named "steven"
by "@steven" or just "steven". This breaks the stated contract for the
Initiatives plugin, but we're planning to deprecate and replace all of
the current data model, so adding some technical debt is OK. And we need
this to unblock beta launch because both MetaGame and SourceCred still
have old-style initiatives.

Test plan: I used this to load SourceCred's Cred (which includes
identities specified in this way) and verified through the explorer UI
that those identities were receiving Cred as expected.