Factor Discourse node and edge helper methods out (#2037)
This commit reorganizes the Discourse helper methods for creating graph
nodes and edges into a helper module called `nodesAndEdges`. This will
create some breathing room for a rewrite of the createGraph module,
without feeling like that module will be over-cluttered. Conceptually,
the exact details of manipulating address structures to create graph
nodes and edges is "below the fold", we'd rather focus on the logic of
weaving the data together into a Graph.

I've moved the unit tests that particularly tested node/edge creation
into the associated module.

Test plan: `yarn test` passes. This is a really simple reorg, that's
amply sufficient.