discourse: mint cred based on trust level
This commit modifies the Discourse plugin so that it sets a weight on
every Like based on the trust level of the user in question. Right now
the weights are hardcoded as follows (we will make them configurable in
the future):

| Trust Level | Weight |
| --- | --- |
| 0 | 0x |
| 1 | 0.1x |
| 2 | 1x |
| 3 | 1.25x |
| 4 | 1.5x |

See the following blog post for details on the trust levels:
https://blog.discourse.org/2018/06/understanding-discourse-trust-levels/

I set the weights very low on levels 0 and 1 because they involve very
little engagement (e.g. don't require making any posts). I set the
weights on 3 and 4 to be only moderately higher than default so as to
keep the democratic vibe.

Test plan: See included unit tests. I'll post a before/after look on the
pull request.

This closes #1863. Thanks to @befitsandpiper for the work that enabled
this PR, namely #2040, #2041, and #2042.