Discord: Remove spam when creating graphs
Currently the Discord plugin emits a warning every time it finds a
"dangling" reaction or mention. These happen pretty often in practice,
and don't indicate that anything is wrong, so it's way too noisy.

This commit simply removes the warnings.

Test plan: Run `sourcecred graph sourcecred/discord` in an instance with
a nontrivial Discord server, and enjoy the cleanliness of the console.