add uuids to ledger events (#2054)
Based on discussions with @wchargin, I believe that having uuids on the
ledger events will be quite valuable. And there's basically no possible
world in which we're sad that we can consistently identify ledger
events.

Test plan: Pretty simple to add, I added a new test that verifies that
we have uuids on the events. And fixed up existing tests.

I wound up making the LedgerLog a transparent type so that I could write
more convenient test code. I don't think there's a harm in letting
people know that the log is an array of events. :) If someone
constructed an invalid event log it'd fail fast due to ledger
validation.