[wip]: frontend shares ledger state
Currently, if one frontend component changes the ledger state, that
doesn't propagate over to the other components, e.g. if you make a new
identity in the admin, then switch to the transfer view, you can't
transfer to the new identity. This commit fixes this by adding a shared
ledger state across all of the admin components.

Test plan: Make a new identity, activate it, and switch to the transfers
pane and transfer to the new identity.