frontend shares ledger state (#2056)
* [wip]: frontend shares ledger state

Currently, if one frontend component changes the ledger state, that
doesn't propagate over to the other components, e.g. if you make a new
identity in the admin, then switch to the transfer view, you can't
transfer to the new identity. This commit fixes this by adding a shared
ledger state across all of the admin components.

Test plan: Make a new identity, activate it, and switch to the transfers
pane and transfer to the new identity.

* convert AdminInner to JSX component

With hooks, react doesn't know then a function is supposed to be a
new component unless the JSX indicates it is (as opposed to function)

test-plan: the admin app should render as expected

* create AdminInner Props flow type

Co-authored-by: Kevin Siegler <17910833+topocount@users.noreply.github.com>