refactor: aliases have descriptions
This commit refactors the ledger Identity type so that aliases are
`{address, description}` objects rather than raw addresses. We add a
description to the alias for roughly the same reason that we added a
description to Graph Nodes in #1191: for UI purposes it's really helpful
to be able to render a human-readable description, and trying to keep
the descriptions and addresses synchronized in an external data
structure is a big pain.

In particular, this will enable the ledger admin UI to display
already-linked aliases in a human-readable fashion.

Since Aliases are usually generated by referencing an existing node in
the graph, creating a description for aliases should not be difficult,
as we can just use the Graph node's description.

Test plan: Unit tests pass and are updated; verify the frontend still
works properly.