aliasSelector fixes
This commit fixes two issues with the AliasSelector:

1. It improperly accesses private properties of the Ledger. This
prompted the change, because I'm refactoring the ledger and that
property is getting renamed.
2. It showed other identities own addresses as potential aliases to bind
to. This only happened if the other identity was already in the graph
(i.e. was not just freshly created in the UI, but was added prior to the
most recent invocation of the `score` command).

Test plan:
- Run `yarn start` and navigate to the ledger admin pane
- Observe that when you open the add alias selector for a user...
  - you do not see any already-bound aliases (existing behavior
  preserved)
  - you do not see any of the other identities as suggestions (new fix)
  - you still see the single unbound alias an option