initial implementation of Creditor nodes
Creditor nodes are utilized to store user-created cred minting actions.
The cannot be deleted, but they are modifiable. In memory they are
stored in the CredNodeStore, which enforces their structure and
heirarchical relationships with other nodes.

Like Initiatives, Creditor nodes are saved to disk in the instance repo.
Unlike Initiatives, they are not meant to be user-modifiable in this
state, but are rather meant to be updated via a web frontend. Diffing
changes is facilitated through use of jsonLog, which outputs each node
on a separate line.

test-plan: unit tests are provided for all methods in the CredNodeStore