update selectedItems control memo to use new aliases
Now that the aliases property on ledger identities includes the
description in addition to the address, we can render it in the selected
items array. This memo activates on any renders where the
currentIdentity Id changes, and removes all selected aliases from the
array that is rendered. If another identity is selected, the new aliases
are added to the array.

test plan: recalculate the graph and cred before testing, then ensure
that:
- existing aliases now render out their pre-added aliases
- selected aliases rerender with the selected identity's alias upon
selected a new identity