refactor: AliasSelector takes id, not identity
Previously, the AliasSelector had a `currentIdentity` property and a
`setCurrentIdentity` setter. It's simpler to have the AliasSelector just
track the identity by its id, and use the ledger as the source of truth
for the identity. This prevents any desync style bugs where the ledger
has the identity in one state, and the component has the identity in a
different one.

Test plan: Play around with the UI. The currently-existing bug still
exists though.