recalculate available aliases on prop change
Because we're relying on  prop data to calculate the available
aliases, we need to recalculate the available aliases for the dropdown
state immediately _after_ linking an alias. If we run `setAliasSearch`
in the combobox  onStateChange reducer, it will recalculate the
available aliases _before_ the updated ledger props are passed back in
and the just-linked alias will still be visible in the dropdown. By
deferring `setAliasSearch()` until after the account's `aliases` have
updated, we are ensuring the dropdown will be synchronized with the
ledger prop.

test plan: ensure no linked aliases are availabled in the dropdown after
linking an alias to an account