Userlike nodes include plugin name in description
This commit updates our plugins so that we include the plugin name in
userlike descriptions. This is helpful because we often have aggregated
views showing all the userlike nodes (e.g. default explorer viz, or when
choosing aliases) and we need to be able to distinguish which platform
the account is on.

Test plan: Updated snapshots; yarn test passes; manual inspection on a
large cred instance with all 3 plugins active.