Fix bugs with GrainAccountOverview
This fixes two issues:
- It had an unnecessary data dep on the CredView, which meant that
adding a new identity and then navigating to the overview would throw an
error, because the new identity wasn't in the CredView yet. This is now
fixed.
- It didn't attach a key on some react nodes, which caused an error in
the console.

Test plan: Add a new identity, then navigate to the overview; it no
longer errors. Also, check the console, and there are no warnings from
React.