update selectedItems control memo to use new aliases (#2061)
* update selectedItems control memo to use new aliases

Now that the aliases property on ledger identities includes the
description in addition to the address, we can render it in the selected
items array. This memo activates on any renders where the
currentIdentity Id changes, and removes all selected aliases from the
array that is rendered. If another identity is selected, the new aliases
are added to the array.

test plan: recalculate the graph and cred before testing, then ensure
that:
- existing aliases now render out their pre-added aliases
- selected aliases rerender with the selected identity's alias upon
selected a new identity

* remove getSelectedItemProps hook from render

this hook enables keyboard selection and deletion of aliases linked to
the active identity. Given That we're rendering the aliases for
informational purposes only, removing this prevents users from deleting
these props from the frontend (even though this action has no effect
on the backend ledger)

test plan: selected aliases cannot be individually focused-on  using
the arrow keys or mouse and moreover, cannot be deleted using the
keyboard delete button or backspace

* strip markdown from string search

descriptions that are markdown links can now be accessed via literal
search, as opposed to typing in a `[` first

test plan: make sure dscourse links can now be returned when typing a
discourse username into the alias combobox

* conditionally render alias selector

conditional rendering will allow us to simplify state management

test-plan: nothing should break in this commit

* remove useMultipleSelection hook

By utilizing props directly to render linked aliases, we don't need any
multiple selection logic. If we ever implement a feature for unlinking
aliases, we can potentially revisit this.

test plan: ensure linked aliases render when selected from the combobox
and when switching between accounts

* move alias search reset into the combobox reducer

The memo is unnecessary now that this is conditionally rendered

test plan: ensure combobox clears after adding a new alias. This removes
the logic that will clear the combobox input value when switching
between accounts

* redirect unlinked alias markdown links to render as spans

Rendering these markdown links as anchor components makes them awkward
to click on.

test plan: ensure available discourse aliases render in plain text that
can be clicked directly

* refactor: AliasSelector takes id, not identity

Previously, the AliasSelector had a `currentIdentity` property and a
`setCurrentIdentity` setter. It's simpler to have the AliasSelector just
track the identity by its id, and use the ledger as the source of truth
for the identity. This prevents any desync style bugs where the ledger
has the identity in one state, and the component has the identity in a
different one.

Test plan: Play around with the UI. The currently-existing bug still
exists though.

* bugfix: prevent incidental form submission of account name

By only wrapping the elements involved in updating the username inside
the form elements we prevent submission-like events from bubbling up and
submitting the form. This is a simpler alternative to preventing
the bubbling using javascript

test plan: ensure user name's aren't submitted to the ledger when
interacting with the alias selector combobox

* recalculate available aliases on prop change

Because we're relying on  prop data to calculate the available
aliases, we need to recalculate the available aliases for the dropdown
state immediately _after_ linking an alias. If we run `setAliasSearch`
in the combobox  onStateChange reducer, it will recalculate the
available aliases _before_ the updated ledger props are passed back in
and the just-linked alias will still be visible in the dropdown. By
deferring `setAliasSearch()` until after the account's `aliases` have
updated, we are ensuring the dropdown will be synchronized with the
ledger prop.

test plan: ensure no linked aliases are availabled in the dropdown after
linking an alias to an account

Co-authored-by: Dandelion Mané <decentralion@dandelion.io>