bugfix: remove resource menu link generator
Currently we are not using react admin resources for any interfaces, so
this logic is unnecessary. Originally, it'd render a react admin
initiatives interface but since it hasn't been merged, this logic only
renders the dummy resource we don't actually want to show

test plan: any instances with the `sourcecred/initiatives` plugin
enabled should _not_ show the dummy resource in the left menu.