Simplify site output
This commit modifies the site command so it no longer "hoists" its
index.html, favicon, and so forth up to the root. Instead, all site data
is inside the site directory, and it symlinks in `sourcecred.json`, and
the `data/`, `config/` and `output/` directories. This generates much
less pollution, and is conducive to having the entire site in a
gitignored directory or in its own side branch.

Note that if you do want to move it to a side branch, you'll need to
either copy the symlink targets explicitly, or dereference the symlinks.

Test plan: I ran the new site command inside a real cred instance. It
produced a working site directroy, as verified by running the admin
command.