Replace BigInt with big-integer for Safari support
This commit swaps out the implementation of our Grain module so that it
uses the big-integer library rather than native BigInts, because BigInts
are not yet supported on Safari. This makes SourceCred usable on iOS and
Safari immediately.

Since we were using native BigInts for computation but not as an at-rest
representation, this is a really easy swap. Actually, the big-integer
library has a `toString` which sensibly just returns the string
implementation, so we could swap to using big integers at rest, and our
serialization/parsing logic would still work, which was what motivated
the string representation to begin with (see #1938 for context, and
also #1936 for a different appraoch considered). However, that would be
a more involved change as it would require re-writing tests. For now,
let's just make the simple fix and start supporting safari.

Test plan: The change is entirely internal to the grain module, which is
well tested; CI passing gives good confidence that this change landed
cleanly. Also, I ran `yarn start` and now Safari can properly load the
instance instead of getting stuck on a loading screen.