jsonLog: use newline-delimited JSON (#2075)
Summary:
Previously, JSON logs, and therefore ledgers, were stored on disk as
JSON array values. This patch converts them to NDJSON values. As a
result, the format is both streaming and concatenative, which are nice
properties to have. (See #2074 for arguments.) Fixes #2074.

As a consequence, the `Ledger` type no longer has an associated
composable parser, because ledgers are not expected to be stored in
other JSON structures, because as they are not valid JSON objects. If we
really do need to embed the contents of a ledger, we can embed its event
log with the public `LedgerLog` representation.

Test Plan:
Unit tests updated and extended to cover the legacy migration case.
Ledger snapshots show that the new format has propagated. Since the CLI
has changed, clone the SourceCred cred instance (`sourcecred/cred`) and
in that instance run `load`, `graph`, `score`, and `site`, and note that
the ledger is unchanged and the site can still read from it. Then, run
`admin`, and in the “Ledger Admin” view click the “save ledger to disk”
button. Note that the file on disk is rewritten to the new format, and
that after refreshing the site still works and shows the right data.

wchargin-branch: jsonlog-ndjson