Ledger: validate distributions
Previously, the ledger did not validate distributions. This meant the
ledger could recieve a bad distribution at runtime, write it to disk,
and then be unable to load from disk (because on load, it tried to parse
the invalid distribution). As of this commit, the ledger checks that the
distribution is parseable before adding it to the event log.

Test plan: Unit tests included.