Use the typesafe numerics
This large commit refactors the codebase to use the new typesafe
numerics introduced in #2082. It "changes the world" rather than doing
isolated refactors, which I regret, but I was fairly deep in before I
realized how messy the commit would wind up being.

That said, this doesn't have much review burden because it consists
entirely of tightening up type declarations, and is extremely unlikely
to introduce regressions. I'm particularly pleased that we now require
that all node and edge weights are finite and non-negative, as this data
comes from plugins, and will now be validated when the plugin graph is
being parsed, which ensures fast and clear failures if a plugin produces
invalid weights.

Test plan: Refactor without runtime changes to code. `yarn test` passes.