Remove nodes/edges debug info (#2098)
The explorer currently shows the # of nodes and edges at the top of the
screen. This was included as a vestigial development artifact, it's
basically a "console.log" to verify the data pipeline is working. We can
now remove it.

Test plan: UI doesn't show this debug info at the top.