Deduplicate `findUserbyUsername` and `findUsername` (#2102)
The Mirror repository had a public method, `findUsername`, which didn't
return the User but just the user's name, and a private method,
`findUserbyUsername`, which gave the full user. Obviously, given the
user, it's trivial to find the username. So I deduplicated this method
and we now export a method `findUser` which can be used in both cases.

Test plan: `yarn test` suffices.