Discourse: connect topics to posts with likes (#2106)
This adds "topicHasLikedPost" edges from topics to posts within that
topic which were liked. Those edges have weight proportional to the
total like weight of the posts in question. This means that the likes
act as a "guide" on where Cred from the topic then flows, which (among
other things) facilitates minting Cred on topics.

Fixes #1896.

Test plan: Added unit tests for the new behavior, and also I manually
tested it on Maker's Cred instance and saw reasonable/expected results.