Update Discourse default Cred flows
This changes the default Cred flows in Discourse. The major ideas are:
- We revamp how Topics flow Cred. Now that we have a nice outbound edge
structure from topics to liked posts (see #2106), it makes more sense to
flow Cred to topics. Posts now have a weight-1 edge to their containing
topic; which makes sense--the post wouldn't exist without a topic to
participate in. However, topics now have a 1/8 edge to each post; this
ensures topics don't act as cred vaccuums when there are no likes, but
will get overpowered by the topicHasLikedPost edges when there are
likes.

I've taken away the direct Cred flow from topics to the author. My
reasoning is that we can flow sufficient Cred to the topic author via
the post-replies edges (which all terminate in the root post) and via
the flow-to-liked-post edges (the root post often gets a lot of likes).

I've made authors flow Cred to posts they author. This is to prevent a
"Cred Black Hole" situation where the author of a post has basically no
outbound weight, resulting in the synthetic loop giving them enormous
Cred.

I also re-balanced the references edge weights, mostly just based on
intuition. There's definitely room for more thought and improvement
there.

Test plan: I've validated these new weights on the MakerDAO Instance,
and ran it by our partners, and they agreed it was a material
improvement in the Cred quality.