Identity: rename "IdentityName" type to "Login" (#2110)
I'm planning to re-organize the Identity module so that we separate the
concept of a "login" (short, url-sanitized string) from a "displayName".

To do so, I'm going to refactor the Identity type to separate the
user-chosen data from the automatically maintained data. But first, as a
prelude cleanup, I rename the IdentityName type to Login in this commit.

Test plan: Existing unit test coverage suffices.