Add the v2 identity type (#2115)
This adds a new version of the Identity type. Key differences:
- All the state which may be freely changed by the owner of the identity
is grouped into an `IdentityDescription`. This will allow clean events
to update this state, as suggested by @hammadj in [a review].
- The identity no longer has an included address--it will instead be
implicit. This removes a class of desync bugs.
- We now support giving the identity some initial aliases rather than
always starting with empty aliases.

Part of #2109.

[a review]: https://github.com/sourcecred/sourcecred/pull/2080#pullrequestreview-463906325

Test plan: `yarn test` passes.