Simplify the identity plugin declaration (#2116)
This reduces the number of node types in the identity plugin declaration
from 4 to 1. We no longer distinguish at the node type level between
different identity subtypes--this is a requirement, because identity
addresses need to be permanent, but the identity type is not permanent, so we
simply have to stop encoding the type in the address.

Also, this whole declaration is going to disappear soon. See #2109 for
details.

While I was here, I updated the test-instance ledger so that it has some
valid aliases, so it was easier to test whether the identities were
still being handled properly by the frontend.

Test plan: `yarn test --full` (since it changes snapshots), also manual
inspection of the test instance frontend to make sure that identities
still appear in the UI.