Ledger: add support for merging identities
This commit adds a new MergeIdentities action to the Ledger, which
allows merging two identities together. One identity is the "base" and
the other is the "target"; the target is absorbed. The base absorbs the
target's balance, paid amount, all of its aliases, and its own innate
address.

This is part of #2109.

Test plan: Thorough unit tests added. Snapshots updated.