no longer include identity type in the address (#2118)
As of #2080, we will allow changing identity types. This means that we
should no longer encode the type in the address, because addresses are
permanent and types are not.

This commit modifies the identity module so we no longer include the
address. However, we already have some identities in production with the
old structure, so I added a warning comment to the effect that you can't
make specific assumptions about the identity address structure.

This builds on #2116.

Test plan: `yarn test --full`.