identity names: support capitalization (#2122)
This changes the identity name module so that we allow names to
canonically include capitalization. To maintain safety, the ledger
ensures that no two names would have the same lowercase
representation--e.g. "Foo" and "foo" are both valid names, but no
instance may contain both at the same time.

This goes against the counsel of @wchargin, because it could create some
confusion, e.g. if we are ever writing output to the filesystem for each
user and need to worry about case sensitivity on the filesystem. But I
think it's a worthy tradeoff to allow users to have more self-expression
in their chosen names. (As an example, GitHub allows case-sensitive
names.)

Test plan: Unit tests updated. New unit tests added to the ledger to
ensure that createIdentity and renameIdentity error if two identities
would have the same name modulo capitalization, and renameIdentity tests
to ensure that its legal to rename an identity to a name that is the
same modulo capitalization.