create server-info.json
server-info.json will be consumed by the frontend to determine whether
or not to show ledger-modifiable views. When an instance is served
statically, the ledger cannot be updated. However, when a local node
service is used, either via the instance's `yarn serve` or the core
`yarn start`, the ledger can be modified and posted back to the file
system.

test plan: the `yarn build` command includes the server-info.json file
in the the build/static folder and `path/to/local/sourcecred.js site`
command add this file to both the site/static folder and the symlinked
./static folder