Add overrides for node-served instances
any site served by node via `yarn admin` within an instance or `yarn
start` in the sourcecred repo will automatically render the
ledger-modifiable routes. This is accomplished through adding overrides
in the express configs that intercept the request for the
server-info.json file and returns true for hasBackend.

test plan: run `yarn start` and `yarn serve` in the sourcecred repo and
directly within an instance. Observe the ledger modifiable routes
rendering in both cases. Make changes to the ledger and ensure the
changes write back to disk, as before.