disable ledger-modifiable views when site is statically served (#2129)
* create server-info.json

server-info.json will be consumed by the frontend to determine whether
or not to show ledger-modifiable views. When an instance is served
statically, the ledger cannot be updated. However, when a local node
service is used, either via the instance's `yarn serve` or the core
`yarn start`, the ledger can be modified and posted back to the file
system.

test plan: the `yarn build` command includes the server-info.json file
in the the build/static folder and `path/to/local/sourcecred.js site`
command add this file to both the site/static folder and the symlinked
./static folder

* frontend utilizes hasBackend bool

The frontend is configure to render the admin, transfer and hidden
special routes when the bool is true. These routes will not be available
in the menu, nor the build, if this flag is false

test plan: admin and transfer, and special transfer panels should not be
in the menu, or accessible via url input when hasBackend is false.
These routes should still be visible and behave as expected when
hasBackend is true, including updating ledger state.

* Add overrides for node-served instances

any site served by node via `yarn admin` within an instance or `yarn
start` in the sourcecred repo will automatically render the
ledger-modifiable routes. This is accomplished through adding overrides
in the express configs that intercept the request for the
server-info.json file and returns true for hasBackend.

test plan: run `yarn start` and `yarn serve` in the sourcecred repo and
directly within an instance. Observe the ledger modifiable routes
rendering in both cases. Make changes to the ledger and ensure the
changes write back to disk, as before.

* remove duplicated express static config

test plan: this should cause no chaange in site behavior.  starting the
dev server should not throw a load error.